requirements.txt
numpy==1.26.4
opencv-contrib-python==4.5.5.62
pillow==10.2.0
